+++ .babelrc
@@ -1,0 +1,4 @@
+{
+  "plugins": ["@babel/plugin-transform-modules-commonjs", "@babel/plugin-proposal-object-rest-spread"]
+}
+
