+++ package.json
@@ -1,12 +1,16 @@
 {
   "devDependencies": {
+    "@babel/core": "^7.0.0-0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.7.7",
+    "@babel/plugin-transform-modules-commonjs": "^7.7.5",
+    "@babel/register": "^7.7.7",
     "@google/earthengine": "^0.1.203",
     "babel-eslint": "^10.0.3",
-    "cypress": "^3.6.1",
+    "cypress": "^3.8.0",
     "eslint": "^6.7.1",
     "eslint-config-google": "^0.14.0",
-    "firebase": "^7.2.3",
-    "firebase-admin": "^8.6.1",
+    "firebase": "^7.6.1",
+    "firebase-admin": "^8.8.0",
     "local-web-server": "^3.0.7"
   },
   "license": "MIT"
